+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/notifications/Notification.scala
@@ -2,6 +2,8 @@ package pl.touk.nussknacker.ui.notifications
 
 import io.circe.generic.JsonCodec
 import io.circe.{Decoder, Encoder}
+import pl.touk.nussknacker.engine.api.deployment.ProcessActionType
+import pl.touk.nussknacker.engine.api.deployment.ProcessActionType.ProcessActionType
 import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.ui.notifications.DataToRefresh.DataToRefresh
 
@@ -13,16 +15,21 @@ import pl.touk.nussknacker.ui.notifications.DataToRefresh.DataToRefresh
                                    toRefresh: List[DataToRefresh])
 
 object Notification {
-  def deploymentFailedNotification(id: String, name: ProcessName, failureMessageOpt: Option[String]): Notification = {
-    Notification(id, Some(name), s"Deployment of ${name.value} failed" + failureMessageOpt.map(" with reason: " + _).getOrElse(""),
+  def actionFailedNotification(id: String, actionType: ProcessActionType, name: ProcessName, failureMessageOpt: Option[String]): Notification = {
+    Notification(id, Some(name), s"${displayableActionName(actionType)} of ${name.value} failed" + failureMessageOpt.map(" with reason: " + _).getOrElse(""),
       Some(NotificationType.error), List(DataToRefresh.state))
   }
 
-  def deploymentFinishedNotification(id: String, name: ProcessName): Notification = {
-    //We don't want to display this notification, not to confuse user, as
-    //deployment may proceed asynchronously (e.g. in streaming-lite)
-    Notification(id, Some(name), s"Deployment finished", None, List(DataToRefresh.versions, DataToRefresh.activity))
+  def actionFinishedNotification(id: String, actionType: ProcessActionType, name: ProcessName): Notification = {
+    // We don't want to display this notification, because user already see that status icon was changed
+    Notification(id, Some(name), s"${displayableActionName(actionType)} finished", None, List(DataToRefresh.versions, DataToRefresh.activity, DataToRefresh.state))
   }
+
+  private def displayableActionName(actionType: ProcessActionType): String =
+    actionType match {
+      case ProcessActionType.Deploy => "Deployment"
+      case ProcessActionType.Cancel => "Cancel"
+    }
 
 }