+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/notifications/NotificationService.scala
@@ -24,11 +24,14 @@ class NotificationServiceImpl(actionRepository: DbProcessActionRepository[DB], d
     val limitBasedOnConfig = now.minusMillis(config.duration.toMillis)
     def maxInstant(instant1: Instant, instant2: Instant) = if (instant1.compareTo(instant2) > 0) instant1 else instant2
     val limit = notificationsAfter.map(maxInstant(_, limitBasedOnConfig)).getOrElse(limitBasedOnConfig)
-    dbioRunner.run(actionRepository.getUserActionsAfter(user, Set(ProcessActionState.Failed, ProcessActionState.Finished), ProcessActionType.Deploy, limit)).map(_.map {
-      case (ProcessActionEntityData(id, _, _, _, _, _, _, ProcessActionState.Finished, _, _, _), processName) =>
-        Notification.deploymentFinishedNotification(id.toString, processName)
-      case (ProcessActionEntityData(id, _, _, _, _, _, _, ProcessActionState.Failed, failureMessageOpt, _, _), processName) =>
-        Notification.deploymentFailedNotification(id.toString, processName, failureMessageOpt)
+    dbioRunner.run(actionRepository.getUserActionsAfter(user,
+      Set(ProcessActionType.Deploy, ProcessActionType.Cancel),
+      Set(ProcessActionState.Failed, ProcessActionState.Finished),
+      limit)).map(_.map {
+      case (ProcessActionEntityData(id, _, _, _, _, _, actionType, ProcessActionState.Finished, _, _, _), processName) =>
+        Notification.actionFinishedNotification(id.toString, actionType, processName)
+      case (ProcessActionEntityData(id, _, _, _, _, _, actionType, ProcessActionState.Failed, failureMessageOpt, _, _), processName) =>
+        Notification.actionFailedNotification(id.toString, actionType, processName, failureMessageOpt)
       case (a, processName) =>
         throw new IllegalStateException(s"Unexpected action returned by query: $a, for scenario: $processName")
     }.toList)