+++ docs/Changelog.md
@@ -17,6 +17,9 @@
 * [#4055](https://github.com/TouK/nussknacker/pull/4055)[#4080](https://github.com/TouK/nussknacker/pull/4080) Removed local state of designer - for HA purpose
 * [#4055](https://github.com/TouK/nussknacker/pull/4055) Performance tweaks for API operations like: process status, deploy, cancel 
 * [#3675](https://github.com/TouK/nussknacker/pull/3675) Improvements: Normalize kafka params name
+* [#4101](https://github.com/TouK/nussknacker/pull/4101) Notifications fixes:
+  * Scenario state wasn't refreshed after deploy/cancel action was successfully finished (was only after failure)
+  * Notification "Deployment of ... failed ..." was presented even for cancel action 
 * [#4100](https://github.com/TouK/nussknacker/pull/4100) "In progress" state is now presented not only inside scenario editor but also on scenarios list
 
 1.8.1 (28 Feb 2023)